requirements.txt
Django==4.1.3
asgiref==3.5.2
django-extensions==3.2.1
django-model-utils==4.3.1
ipdb==0.13.9
sqlparse==0.4.3